\name{swift-class}
\docType{class}
\alias{swift-class}

\title{Class "swift" }
\description{A (virtual) class that extends \code{brob} and
  \code{glub} objects}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "swift" in the signature.
}
\author{Robin K. S. Hankin}
\seealso{
  \code{\link{brob-class}},\code{\link{glub-class}}
}
\keyword{classes}
